# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/transcribe": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload an audio file for transcription and optional summarization",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transcription"
                ],
                "summary": "Upload and transcribe audio file",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Audio file (WAV, MP3, FLAC, M4A)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Enable AI summarization of the transcript",
                        "name": "enable_summarization",
                        "in": "formData"
                    },
                    {
                        "enum": [
                            "high",
                            "medium",
                            "low"
                        ],
                        "type": "string",
                        "description": "Transcription quality (high, medium, low)",
                        "name": "quality",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Language code for transcription (auto-detect if not specified)",
                        "name": "language",
                        "in": "formData"
                    },
                    {
                        "type": "boolean",
                        "description": "Enable stereo channel separation for speaker diarization",
                        "name": "channel_separation",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transcription completed successfully",
                        "schema": {
                            "$ref": "#/definitions/models.TranscriptionResponse"
                        }
                    },
                    "202": {
                        "description": "Transcription queued for processing",
                        "schema": {
                            "$ref": "#/definitions/models.AsyncResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or file format",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "File too large",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "415": {
                        "description": "Unsupported media type",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transcribe/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a previously completed transcription by its session ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transcription"
                ],
                "summary": "Get transcription by session ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID for the transcription",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transcription found",
                        "schema": {
                            "$ref": "#/definitions/models.TranscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid session ID",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Transcription not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.AsyncResponse": {
            "description": "Response for asynchronous operations that are queued for processing",
            "type": "object",
            "properties": {
                "check_url": {
                    "type": "string",
                    "example": "/api/v1/transcribe/job_12345"
                },
                "estimated_eta": {
                    "type": "string",
                    "example": "30s"
                },
                "job_id": {
                    "type": "string",
                    "example": "job_12345"
                },
                "message": {
                    "type": "string",
                    "example": "Audio file queued for processing"
                },
                "request_id": {
                    "type": "string",
                    "example": "req_12345"
                },
                "status": {
                    "type": "string",
                    "example": "queued"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-10-15T14:30:45Z"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "models.Metadata": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "audio duration in seconds",
                    "type": "number"
                },
                "model_versions": {
                    "description": "AI model versions",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ModelVersions"
                        }
                    ]
                },
                "processed_at": {
                    "description": "timestamp of processing",
                    "type": "string"
                },
                "processing_time": {
                    "description": "processing time in seconds",
                    "type": "number"
                }
            }
        },
        "models.ModelVersions": {
            "type": "object",
            "properties": {
                "llama": {
                    "description": "e.g., \"7b\"",
                    "type": "string"
                },
                "whisper": {
                    "description": "e.g., \"large-v3\"",
                    "type": "string"
                }
            }
        },
        "models.Summary": {
            "description": "AI-generated summary of the transcribed audio content",
            "type": "object",
            "properties": {
                "key_points": {
                    "description": "extracted key points",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Project timeline discussion\"",
                        " \"Budget allocation\"",
                        " \"Team assignments\"]"
                    ]
                },
                "sentiment": {
                    "description": "\"positive\", \"neutral\", \"negative\"",
                    "type": "string",
                    "example": "positive"
                },
                "text": {
                    "description": "main summary text",
                    "type": "string",
                    "example": "This conversation discusses project planning and resource allocation."
                }
            }
        },
        "models.Transcript": {
            "description": "Complete transcription result containing all segments",
            "type": "object",
            "properties": {
                "segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TranscriptSegment"
                    }
                }
            }
        },
        "models.TranscriptSegment": {
            "description": "A single segment of transcribed audio with speaker information and timing",
            "type": "object",
            "properties": {
                "end_time": {
                    "description": "in seconds",
                    "type": "number",
                    "example": 3.2
                },
                "speaker": {
                    "description": "\"left\" or \"right\"",
                    "type": "string",
                    "example": "left"
                },
                "start_time": {
                    "description": "in seconds",
                    "type": "number",
                    "example": 0.5
                },
                "text": {
                    "description": "transcribed text",
                    "type": "string",
                    "example": "Hello, this is a test."
                }
            }
        },
        "models.TranscriptionResponse": {
            "type": "object",
            "properties": {
                "metadata": {
                    "$ref": "#/definitions/models.Metadata"
                },
                "summary": {
                    "$ref": "#/definitions/models.Summary"
                },
                "transcript": {
                    "$ref": "#/definitions/models.Transcript"
                }
            }
        },
...
    }
}